Add list of comma separated properties in constants
